Basic authentication interface
Addition of the "login.php" file to present a simple authentication interface for users.

The current interface is crude, and will be refined after all the basic system functionality is implemented.

This commit also contains some a simple "index.php" to test authentication. This file will now be expanded to implement the basic system functionality.