Basic Initial Interface
Modification of the "index.php" file to enable access to the main functionality that will be implemented on the system.

For now, the index is a really simple interface that will be updated after the initial implementation of all system functionality.

The "login.php" was also modified to include a cookie field to identify admin users, which can also manage user and room information. This is NOT safe, and will ONLY be used during initial system development.