User Listing Interface
Added a basic interface to list all users on the system and enable their addition, update and removal.

This interface is implemented on the "list-users.php" file. It is still crude and lacks features such as user querying and sorting. Such features are planned to be implemented after all system basic functionality is available.

The file "database.php" also had additions to enable the retrieval of user information for the above interface.