User Management Interface
This commit adds the "edit-users.php" file, which contains the interface to add, update and remove users from the system.

This interface is accessed from the user listing (list-users.php). The interface will adapt itself depending on the requested operation. More specifically, it will present blank form fields if a new user should be added, or will load user information from the database if an existing user should be updated or removed.

This interface is still crude (there are no confirmation requests for removal and passwords are shown clear). These are known issues that will be fixes as soon as the remaining basic interfaces have their functionality implemented.

The database.php and list-users.php files were also updated to include the necessary routines to support operations from the above interface. In the case of database operations, methods were added for each operation that can be executed with user data.