Rooms Listing and Management
This commit adds the interfaces for listing and managing rooms available for reservation. These are implemented, respectively, in files "list-rooms.php" and "edit-rooms.php".

The code from these interfaces are very similar to those employed in user management. As such, these interfaces are also crude. They will be polished after the functionality of room reservation is implemented in the system.

This commit also include modifications on the "database.php" to implement the required opperations to manage and query information about rooms.