Initial Reservation Listing Interface
This commit introduces an interface to visualize the status of reservations for all rooms in a certain day. It is implemented in the file "list-reservations.php".

This is an initial version of the interface, that allows the user to select a date through an HTML5 date input type. Upon selection, the system recovers the list of rooms and their respective reservations for that day. The interface then presents a grid in which occupied times are marked with an "X" and free times with a "O". The interface will also serve as access point for users to create or modify their room reservations.

This is an initial and crude version of this interface. As soon as the functionality to manage user reservations is added to the system, all interfaces will undergo a refinement process.