Reservation Management Interface
This commit introduces the interface employed to manage reservations. It is implemented in the "edit-reservations.php" file.

This interface requires no user input because it acquires all information to perform room reservations from the listing interface from the "list-reservations.php" file. It only requires the user to confirm if the desired opperations should be carred out. The code of the interface is also responsible to check do not break the system rules (such as more than one room reserved by a user at the same time, or if the user removing a reservation is the same that made it).

This version of this interface is extremely simple and can be further improoved. The implementation of this interface completes the work on the basic functionalities of the system. The following commints will be all related to the improvement of the developed interfaces in order to improve user experience with the system.