AJAX Interface for Reservation Management
This commit changes the behavior of the reservation management interface to use XHR.

The main javascript functions used to request and cancel reservations through AJAX calls are implemented in the "reservations.js" file. They are "RequrestReservation" and "CancelReservation" respectively. Overall, both functions create an asynchronous call to the "edit-reservations.php", which will handle the actual reservation process.

The "edit-reservations.php" script was modified to handle POST parameters instead of the GET used in previous versions. Also, the HTML returned by this script also employs javascript calls in its user inputs to confirm reservation operations.

The "list-reservations.php" script was modified to employ javascript calls to request and cancel room reservations. Previously, these functions were implemented as normal page links to the "edit-reservations.php" script. A new div was also included in the HTTP code to hold the results obtained from the AJAX calls to inform clients about reservations status.

Additionally, the "jquery.js" file was included in the system. It contains the latest release (3.0.0) of the jQuery library, employed to ease the implementation of various javascript functions in the system.

The next step in the system implementation will be to further employ XHR in the reservation interface, more specifically in the of rooms availability per day.