Minor Code Cleaning on Reservation Management
There were some unnecessary code repetition on the "edit-reservations.php" on the section that manages room reservations. These were eliminated.

Also, an unnecessary parameter were passed to "edit-reservations.php" when a reservation was requested. This parameter was removed from the script calls in the "list-reservations.php" and "reservations.js" files.