Implementation of XHR on the Reservation Interface
This commit implements the use of XHR in the entire room reservation interface.

A new file, "reservations.php", was added as an entry point to the room reservation interface. It contains a field to receive the reservation date. This field employs a datepicker UI element from the jQuery library instead of the previously used date HTML5 input element. This change was made due to the higher compatibility of the jQuery UI element with browsers. The form is also preconfigured with the current date as default.

The listing of room availability is automatically loaded through XHR after the page is loaded and also after modifications on the reservation date field. The reservation management within the availability grid still employs the XHR methods implemented in the previous commit.

The "reservations.js" file was updated to include the necessary functions to request the room availability grid through XHR. Also, a function was added to initialize the datepicker UI element.

As result of the implementation of XHR, the "list-reservations.php" file was updated to remove unnecessary HTML code (which is now handled in "reservations.php").

The next step on the system development will be to integrate all system interfaces (user and room management, and reservations) employing XHR.