XHR Interface
The first goal of this commit is to introduce XHR to the main system interface. The "index.php" script was updated to load the remaining system interfaces through XHR instead to direct links to php scripts.

A new file named "main.js" was introduced. This file contains all the javascript code related to the XHR requests used in the main system interface.

The second goal of this commit is to update the user management interface to support XHR through the main system interface. To that end, the "list-users.php" and "edit-users.php" scripts were updated to employ javascript XHR functions instead of direct links.

The "main.js" file also contains the javascript code required by the user management interface through XHR. Functions were included to asynchronously load the list of registered users, the user update interface, and also the POST methods to submit changes to user information.

The next step in the implementation will be to update the room management interfaces to also employ XHR methods.