XHR Interface, Part 2
This commit finishes the implementation of XHR throughout the system interface, now including the room management interface.

The scripts "list-rooms.php" and "edit-rooms.php" were updated to employ javascript XHR calls instead of direct links to php scripts. These calls including interface loading to display the list of rooms or the room information updating. It also includes the submission of forms through POST methods to manage room information.

The commit also include changes in the "main.js" javascript file to support the XHR methods require by the room management interface. This commit contains some code repetition among the methods employed by the room and user management interfaces. This is a known issue that will be dealt with in a future commit.

This commit concludes the implementation of the system interface with the use of XHR. The next step in the development process will involve the styling of the interface to improve its design and look.