Interface Styling with CSS
This commit adds a simple stylesheet to almost all elements of the interface. The styles are stored the file "style.css".

The remaining php scripts were also modified to include the proper attributes in the interface elements to enable the styles defined in the aforementioned file. More information about the changes can be found on the GitHub patch viewer.

The only exception to this commit is the interface to manage room reservations. This interface requires additional styling efforts, which will be implemented in the next commit.