Reservation Interface Styling with CSS
This commit introduces the styling of the reservation grid presented by the "list-reservations.php" file.

The symbols employed to represent busy or free room slots were changed. Also, background colors were introduced to present a more intuitive view of room occupation. A descriptive text was also included to explain the functionality of the grid.

The next step on the system development will be to introduce minor refinements on the interface and code to polish the system.