Introduction of Alert Messages
Many of the system responses to user interaction are simple notifications indicating that an action has been completed successfully. For example, the registration of a reservation or an update in user information.

To improve the interface intuitiveness, these messages are now handled by alerts shown in the main interface. Basically, these alerts are shown in a div declared in the "index.php" file, which is updated using javascript XHR callbacks from the remaining php scripts.

This commit changes the "main.js" and "reservation.js" files to implement the necessary functions to handle the alert messages (hiding and showing the alert div when necessary). They also implement additional callback functions for XHR to handle interface changes depending on the type of alerts being shown.