Additional Data Consistency Checks
Usernames should be unique in the system because they are used for authentication. Similarly, room numbers should be unique because they are used to uniquely identify available rooms for reservations.

User and room management interfaces did not check for uniqueness when adding or updating records to avoid duplicates of the aforementioned fields. This commit implements these checks.

Two methods were added to the "database.php" script to enable checking if an username or a room number is already registered in the system.

The scripts "edit-users.php" and "edit-rooms.php" were also modified to check if an username or a room number is not already registered. If such a situation occurs, it will alert the user with a banner. The user will have the opportunity to change the value to be registered without loosing already entered information in the management interfaces.