Logout Option
This commit adds the logout option to the main menu in the "index.php" script. It also introduces the "logout.php" script that takes care of actual user logout procedures.

Some minor modifications were also made to the interface to better present information on the "list-users.php" and "list-rooms.php" scripts.