Password Hashing
This commit introduces the use of SHA256 hashing to store user passwords.

The "login.php" script was modified to compare the SHA256 hash of user input passwords instead of plaintext. Also, the "edit-user.php" script was modified to save passwords with SHA256 hashes instead of plaintext.

Also, the interface to edit user information now displays a blank form in the password field. The password information of the user will be updated ONLY if a new password is typed in this field. If the field is left blank, only the remaining user information is updated.

This commit implements most of the features planned for the room reservation system. The next step of development will include the code documentation in order to ease further development.