Confirmation for Removal
The interfaces to manage rooms and users did not ask for confirmation when the option to remove a registered entity was selected. Now, each time a room or user is removed, an alert is shown asking if the operation should really be executed. Additionally, the system will not allow the first administrative user (id 1 in the database) to be removed.

This commit introduces changes in the "edit-rooms.php" and "edit-users.php" to add the confirmation messages to the XHR callback of the interface. It also introduces changes to the "main.js" script to enable the presentation of alert messages asking for confirmation in the room and user management interfaces.