Configuration File
This commit introduces the "config.php" file, which concentrates a set of configuration variables used throughout the system, which may be edited depending on the needs of the system administrator.

These variables are mainly used in the "database.php" file to set up database connection and in the "list-reservations.php" file to specify available hour slots for rooms. More information about each variable will be added in the script comments.

The next commits will contain the code documentation to clarify the system implementation.